It's working! - Completely redid models; Autogenerated models against Usda response to make my life easier; Updated mapper and output based on model updates

+++ Models/FoodDto.cs
@@ -2,7 +2,7 @@
 
 public class FoodDto
 {
-    public int FdcId { get; set; }
+    public long FdcId { get; set; }
     public string Description { get; set; }
     public List<FoodNutrientDto> FoodNutrients { get; set; }
     public List<FoodMeasureDto> FoodMeasures { get; set; }
